==> Code/01. Libraries.R <==
###################################
### Cargar librerías necesarias ###
###################################

# 📦 Manipulación y limpieza de datos
library(tidyverse)
library(janitor)    
library(lubridate)

# 📥 Importación y exportación de datos
library(readxl)    
library(rio)       
library(writexl)   

# 🌍 Datos espaciales
library(sp)        
library(sf)        
library(gstat)     
library(automap)

# 🎨 Visualización de datos
library(viridis)
library(gifski) 
library(gridExtra)
